Revert "Add helmfile/terraform runner with hierarchical configuration support (#44)"
This reverts commit 5d7342bdceb101eb9d1fff04efff91d9f4e01ed0.
