+++ build_scripts/Dockerfile
@@ -73,7 +73,6 @@ RUN curl -sSL https://github.com/databus23/helm-diff/releases/download/v${HELM_D
 
 USER root
 RUN HELM_HOME=/home/ops/.helm helm plugin install https://github.com/futuresimple/helm-secrets
-RUN HELM_HOME=/home/ops/.helm helm plugin install https://github.com/rimusz/helm-tiller
 RUN chown -R ops:ops /home/ops/.helm/plugins
 
 COPY --from=compile-image /azure-cli /home/ops/.local/azure-cli
