+++ build_scripts/docker_push.sh
@@ -2,5 +2,5 @@
 set -e
 
 echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin
-docker tag ops adobe/ops-cli:1.0
-docker push adobe/ops-cli:1.0
+docker tag ops adobe/ops-cli:0.36
+docker push adobe/ops-cli:0.36
