--- examples/features/terraform-hierarchical/.opsconfig.yaml
@@ -1,9 +1,0 @@
----
-compositions_order:
-  terraform:
-    - account
-    - network
-    - cluster
-    - spinnaker
-  helmfile:
-    - helmfiles
