--- examples/features/terraform-hierarchical/compositions/terraform/cluster/main.tf
@@ -1,10 +1,0 @@
-variable "config" {}
-
-module "cluster" {
-  source = "../../../modules/cluster"
-  config = var.config
-}
-
-output "cluster_name" {
-  value = var.config.cluster.name
-}
