--- examples/features/terraform-hierarchical/compositions/terraform/network/main.tf
@@ -1,6 +1,0 @@
-variable "config" {}
-
-module "network" {
-  source = "../../../modules/network"
-  config = var.config
-}
