--- examples/features/terraform-hierarchical/config/env=dev/cluster=cluster1/conf.yaml
@@ -1,2 +1,0 @@
-cluster:
-  name: cluster1
