--- examples/features/terraform-hierarchical/config/env=dev/cluster=cluster2/conf.yaml
@@ -1,2 +1,0 @@
-cluster:
-  name: cluster2
