--- examples/features/terraform-hierarchical/modules/cluster/main.tf
@@ -1,5 +1,0 @@
-variable "config" {}
-
-output "cluster_name" {
-  value = var.config.cluster.name
-}
