+++ src/ops/__init__.py
@@ -58,4 +58,3 @@ class Executor(object):
 
 class OpsException(Exception):
     pass
-
