+++ src/ops/cli/terraform.py
@@ -1,21 +1,24 @@
-# Copyright 2019 Adobe. All rights reserved.
-# This file is licensed to you under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License. You may obtain a copy
-# of the License at http://www.apache.org/licenses/LICENSE-2.0
+#Copyright 2019 Adobe. All rights reserved.
+#This file is licensed to you under the Apache License, Version 2.0 (the "License");
+#you may not use this file except in compliance with the License. You may obtain a copy
+#of the License at http://www.apache.org/licenses/LICENSE-2.0
 
-# Unless required by applicable law or agreed to in writing, software distributed under
-# the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
-# OF ANY KIND, either express or implied. See the License for the specific language
-# governing permissions and limitations under the License.
+#Unless required by applicable law or agreed to in writing, software distributed under
+#the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+#OF ANY KIND, either express or implied. See the License for the specific language
+#governing permissions and limitations under the License.
 
 import os
-import hashlib
-import logging
-from ops.cli.parser import SubParserConfig
-from ops.terraform.terraform_cmd_generator import TerraformCommandGenerator
-from ops.hierarchical.composition_config_generator import TerraformConfigGenerator
+import re
+import shutil
 
-logger = logging.getLogger(__name__)
+from jinja2 import Environment, FileSystemLoader
+from jinja2.runtime import DebugUndefined, StrictUndefined
+
+from ops.cli.parser import SubParserConfig
+from subprocess import Popen, PIPE
+
+from . import aws, err, display
 
 
 class TerraformParserConfig(SubParserConfig):
@@ -26,47 +29,24 @@ class TerraformParserConfig(SubParserConfig):
         return 'Wrap common terraform tasks with full templated configuration support'
 
     def configure(self, parser):
-        parser.add_argument('subcommand',
-                            help='apply | console | destroy | import | output | plan | refresh | show | taint | template | untaint | validate',
-                            type=str)
+        parser.add_argument('subcommand', help='apply | console | destroy | import | output | plan | refresh | show | taint | template | untaint | validate', type=str)
         parser.add_argument('--var', help='the output var to show', type=str, default='')
-        parser.add_argument('--module',
-                            help='for use with "taint", "untaint" and "import". The module to use. e.g.: vpc', type=str)
-        parser.add_argument('--resource',
-                            help='for use with "taint", "untaint" and "import". The resource to target. e.g.: aws_instance.nat',
-                            type=str)
-        parser.add_argument('--name',
-                            help='for use with "import". The name or ID of the imported resource. e.g.: i-abcd1234',
-                            type=str)
-        parser.add_argument('--plan', help='for use with "show", show the plan instead of the statefile',
-                            action='store_true')
-        parser.add_argument('--state-location', help='control how the remote states are used',
-                            choices=['local', 'remote', 'any'], default='any', type=str)
-        parser.add_argument('--force-copy',
-                            help='for use with "plan" to do force state change automatically during init phase',
-                            action='store_true')
-        parser.add_argument('--template-location',
-                            help='for use with "template". The folder where to save the tf files, without showing',
-                            type=str)
-        parser.add_argument('--skip-refresh', help='for use with "plan". Skip refresh of statefile',
-                            action='store_false', dest='do_refresh')
+        parser.add_argument('--module', help='for use with "taint", "untaint" and "import". The module to use. e.g.: vpc', type=str)
+        parser.add_argument('--resource', help='for use with "taint", "untaint" and "import". The resource to target. e.g.: aws_instance.nat', type=str)
+        parser.add_argument('--name', help='for use with "import". The name or ID of the imported resource. e.g.: i-abcd1234', type=str)
+        parser.add_argument('--plan', help='for use with "show", show the plan instead of the statefile', action='store_true')
+        parser.add_argument('-i', '--interactive', help='for use with "apply", use the new interactive apply workflow introduced in TF 0.11.0', action='store_true')
+        parser.add_argument('--state-location', help='control how the remote states are used', choices=[ 'local', 'remote', 'any'], default='any', type=str)
+        parser.add_argument('--force-copy', help='for use with "plan" to do force state change automatically during init phase', action='store_true')
+        parser.add_argument('--template-location', help='for use with "template". The folder where to save the tf files, without showing', type=str)
+        parser.add_argument('--skip-refresh', help='for use with "plan". Skip refresh of statefile', action='store_false', dest='do_refresh')
         parser.set_defaults(do_refresh=True)
-        parser.add_argument('--raw-output',
-                            help='for use with "plan". Show raw plan output without piping through terraform landscape - '
-                                 'https://github.com/coinbase/terraform-landscape (if terraform landscape is not enabled in opsconfig.yaml '
-                                 'this will have no impact)', action='store_true',
+        parser.add_argument('--raw-output', help='for use with "plan". Show raw plan output without piping through terraform landscape - '
+                                                 'https://github.com/coinbase/terraform-landscape (if terraform landscape is not enabled in opsconfig.yaml '
+                                                 'this will have no impact)', action='store_true',
                             dest='raw_plan_output')
         parser.set_defaults(raw_plan_output=False)
-        parser.add_argument('--path-name',
-                            help='in case multiple terraform paths are defined, this allows to specify which one to use when running terraform',
-                            type=str)
-        parser.add_argument('--terraform-path', type=str, default=None, help='Path to terraform files')
-        parser.add_argument('--skip-plan',
-                            help='for use with "apply"; runs terraform apply without running a plan first',
-                            action='store_true')
-        parser.add_argument('--auto-approve',
-                            help='for use with "apply". Proceeds with the apply without waiting for user confirmation.',
-                            action='store_true')
+        parser.add_argument('--path-name', help='in case multiple terraform paths are defined, this allows to specify which one to use when running terraform', type=str)
         parser.add_argument('terraform_args', type=str, nargs='*', help='Extra terraform args')
 
         return parser
@@ -74,12 +54,18 @@ class TerraformParserConfig(SubParserConfig):
     def get_epilog(self):
         return '''
     Examples:
-        # Create/update a new cluster with Terraform
+        # Create a new cluster with Terraform
         ops clusters/qe1.yaml terraform plan
         ops clusters/qe1.yaml terraform apply
 
-        # Run Terraform apply without running a plan first
-        ops clusters/qe1.yaml terraform apply --skip-plan
+        # Update an existing cluster
+        ops clusters/qe1.yaml terraform plan
+        ops clusters/qe1.yaml terraform apply
+
+        # Run Terraform apply without running a plan first, runs in an interactive mode
+        ops clusters/qe1.yaml terraform apply --interactive
+        # or
+        ops clusters/qe1.yaml terraform apply -i
 
         # Get rid of a cluster and all of its components
         ops clusters/qe1.yaml terraform destroy
@@ -119,67 +105,431 @@ class TerraformParserConfig(SubParserConfig):
 
         # Specify which terraform path to use
         ops clusters/qe1.yaml terraform plan --path-name terraformFolder1
-        
-        # Run terraform v2 integration
-        ops data/env=dev/region=va6/project=ee/cluster=experiments terraform plan
         '''
 
 
 class TerraformRunner(object):
-    def __init__(self, root_dir, cluster_config_path, cluster_config, inventory_generator, ops_config, template,
-                 execute):
-        self.cluster_config_path = cluster_config_path
+    def __init__(self, root_dir, cluster_config, inventory_generator, ops_config, template):
         self.cluster_config = cluster_config
         self.root_dir = root_dir
         self.inventory_generator = inventory_generator
         self.ops_config = ops_config
         self.template = template
-        self.execute = execute
 
     def run(self, args):
-        if os.path.isdir(self.cluster_config_path):
-            return self.run_v2_integration(args)
-        else:
-            return self.run_v1_integration(args)
 
-    def run_v1_integration(self, args):
-        return self.run_composition(args, self.cluster_config)
-
-    def run_composition(self, args, config):
-        generator = TerraformCommandGenerator(self.root_dir,
-                                              config,
-                                              self.inventory_generator,
-                                              self.ops_config,
-                                              self.template)
-        return generator.generate(args)
-
-    def run_v2_integration(self, args):
-        logging.basicConfig(level=logging.INFO)
-        config_path = os.path.join(self.cluster_config_path, '')
-        terraform_path = '../ee-k8s-infra/' if args.terraform_path is None else os.path.join(args.terraform_path, '')
-        terraform_path = '{}compositions/terraform/'.format(terraform_path)
-        composition_order = self.cluster_config.ops_config.config["compositions_order"]["terraform"]
-
-        tf_config_generator = TerraformConfigGenerator(composition_order)
-        reverse_order = "destroy" == args.subcommand
-        compositions = tf_config_generator.get_sorted_compositions(config_path, reverse=reverse_order)
-        if len(compositions) == 0:
-            raise Exception("No terraform compositions were detected for it in %s.", self, config_path)
-
-        for composition in compositions:
-            logger.info("Running composition: %s", composition)
-            tf_config_generator.generate_files(config_path, terraform_path, composition)
-            command = self.run_v2_composition(args, terraform_path, composition)
-            ret = self.execute(command)
-            if ret != 0:
-                logger.error("Command finished with nonzero exit code. Will skip remaining compositions.")
-                return ret
-        return 0
-
-    def run_v2_composition(self, args, terraform_path, composition):
+        self.selected_terraform_path = args.path_name
+        self.set_current_working_dir()
+        current_terraform_version = self.check_terraform_version()
         config = self.cluster_config
-        config['terraform'] = {}
-        config['terraform']["path"] = "{}{}".format(terraform_path, composition)
-        config['terraform']["variables_file"] = "variables.tfvars.json"
-        config['cluster'] = hashlib.md5(self.cluster_config_path).hexdigest()[:6]
-        return self.run_composition(args, config)
+
+        current_terraform_version_major = int(current_terraform_version.split('.')[1])
+        if 'enable_consul_remote_state' in config['terraform']:
+            terraform_remote_state = config['terraform']['enable_consul_remote_state']
+        elif config['terraform'].get('state', {'type': None}).get('type') == 's3':
+            terraform_remote_state = 'true'
+        else:
+            terraform_remote_state = 'false'
+
+        terraform_config = config.get('terraform', {})
+        terraform_path = self.get_terraform_path()
+        generate_module_templates = False
+
+        plan_variables = terraform_config.get('vars', {})
+        plan_variables['cluster'] = config['cluster']
+        if self.cluster_config.has_ssh_keys:
+            plan_variables['has_ssh_keys'] = True
+            plan_variables['cluster_ssh_pubkey_file'] = self.cluster_config.cluster_ssh_pubkey_file
+            plan_variables['cluster_ssh_prvkey_file'] = self.cluster_config.cluster_ssh_prvkey_file
+        if terraform_config.get('boto_profile'):
+            self.add_profile_vars(plan_variables, terraform_config.get('boto_profile'))
+
+        vars = ''
+        for key, val in plan_variables.items():
+            vars += " -var '%s=%s' " % (key, val)
+
+        state_file = 'terraform.{cluster}.tfstate'.format(cluster=config['cluster'])
+        plan_file = 'terraform.{cluster}.plan'.format(cluster=config['cluster'])
+        landscape = ''
+
+        if current_terraform_version_major >= 9:
+            if args.force_copy:
+                terraform_init_command = 'terraform init -force-copy && '
+            else:
+                terraform_init_command = 'terraform init && '
+            # regarding state location we give priority to the cli parameter
+            if args.state_location == 'remote':
+                state_argument = ''
+                state_out_argument = ''
+            elif args.state_location == 'local':
+                state_argument = "-state={state_file}".format(
+                    state_file = state_file
+                )
+                state_out_argument = "-state-out={state_file}".format(
+                    state_file = state_file
+                )
+            else:
+                # no cli parameter, decide based on config file
+                if terraform_remote_state == 'true':
+                    state_argument = ''
+                    state_out_argument = ''
+                else:
+                    state_argument = "-state={state_file}".format(
+                        state_file=state_file
+                    )
+                    state_out_argument = "-state-out={state_file}".format(
+                        state_file=state_file
+                    )
+        else:
+            state_argument = "-state={state_file}".format(
+                state_file=state_file
+            )
+            state_out_argument = "-state-out={state_file}".format(
+                state_file=state_file
+            )
+            terraform_init_command = ''
+
+        if args.subcommand == 'template':
+            if args.template_location:
+                self.copy_static_files(args.template_location, terraform_path)
+                self.write_module_templates(args.template_location)
+                self.write_var_file(os.path.join(args.template_location, terraform_path), plan_variables)
+            else:
+                for original, fname, contents in self.get_templated_files():
+                    display("# %s -> %s" % (original, fname), color="green")
+                    display("# --------------", color="green")
+                    display(contents)
+            return
+
+        if args.subcommand == 'plan':
+            generate_module_templates = True
+            terraform_refresh_command = remove_local_cache = ''
+            if args.do_refresh:
+                terraform_refresh_command = "terraform refresh -input=false {vars} {state_argument} && ".format(vars=vars, state_argument=state_argument)
+
+            if self.ops_config['terraform.landscape'] and not args.raw_plan_output:
+                landscape = '| landscape'
+
+            if self.ops_config['terraform.remove_local_cache']:
+                remove_local_cache = 'rm -rf .terraform && '
+
+            cmd = "cd {root_dir}/{terraform_path} && " \
+                  "{remove_local_cache}" \
+                  "terraform get -update && " \
+                  "{terraform_init_command}" \
+                  "{terraform_refresh_command}" \
+                  "terraform plan " \
+                  "-out={plan_file} -refresh=false -input=false {vars} {state_argument}".format(
+                    root_dir=self.root_dir,
+                    terraform_path=terraform_path,
+                    terraform_init_command=terraform_init_command,
+                    vars=vars,
+                    state_argument=state_argument,
+                    plan_file=plan_file,
+                    terraform_refresh_command=terraform_refresh_command,
+                    remove_local_cache=remove_local_cache
+            )
+
+        elif args.subcommand == 'apply':
+            # the following is to have auxiliary rendered/templated files like cloudinit.yaml
+            # that also needs templating. Without it, plan works but apply does not for this kind of files
+            # todo maybe this deserves a better implementation later
+            generate_module_templates = True
+
+            self.inventory_generator.clear_cache()
+            if args.interactive:
+                cmd = "cd {root_dir}/{terraform_path} && {terraform_init_command}" \
+                      "rm -f {plan_file} && terraform apply {vars}" \
+                      "-refresh=true {state_argument}".format(
+                        plan_file=plan_file,
+                        root_dir=self.root_dir,
+                        state_argument=state_argument,
+                        terraform_init_command=terraform_init_command,
+                        terraform_path=terraform_path,
+                        vars=vars,
+                      )
+            else:
+                cmd = "cd {root_dir}/{terraform_path} && " \
+                      "terraform apply " \
+                      "-refresh=true {state_out_argument} {plan_file}; code=$?; rm {plan_file}; exit $code".format(
+                        plan_file=plan_file,
+                        root_dir=self.root_dir,
+                        state_out_argument=state_out_argument,
+                        terraform_path=terraform_path,
+                        vars=vars,
+                      )
+
+        elif args.subcommand == 'destroy':
+            generate_module_templates = True
+            remove_local_cache = ''
+
+            if self.ops_config['terraform.remove_local_cache']:
+                remove_local_cache = 'rm -rf .terraform && '
+            cmd = "cd {root_dir}/{terraform_path} && " \
+                  "{remove_local_cache}" \
+                  "{terraform_init_command}" \
+                  "terraform plan -destroy " \
+                  "-refresh=true {vars} {state_argument} && " \
+                  "terraform destroy {vars} {state_argument} -refresh=true".format(
+                    root_dir=self.root_dir,
+                    terraform_path=terraform_path,
+                    vars=vars,
+                    state_argument=state_argument,
+                    terraform_init_command=terraform_init_command,
+                    remove_local_cache=remove_local_cache
+            )
+        elif args.subcommand == 'output':
+            cmd = "cd {root_dir}/{terraform_path} && " \
+                  "terraform output {state_argument} {output}".format(
+                    root_dir=self.root_dir,
+                    terraform_path=terraform_path,
+                    output=args.var,
+                    state_argument=state_argument
+            )
+        elif args.subcommand == 'refresh':
+            generate_module_templates = True
+            cmd = "cd {root_dir}/{terraform_path} && " \
+                  "terraform get -update && " \
+                  "terraform refresh {state_argument} {vars}".format(
+                    root_dir=self.root_dir,
+                    terraform_path=terraform_path,
+                    vars=vars,
+                    state_argument=state_argument
+            )
+        elif args.subcommand == 'taint' or args.subcommand == 'untaint':
+            cmd = "cd {root_dir}/{terraform_path} && " \
+                  "{terraform_init_command}" \
+                  "terraform {command} {state_argument} -module={module} {resource}".format(
+                    root_dir=self.root_dir,
+                    command=args.subcommand,
+                    terraform_path=terraform_path,
+                    resource=args.resource,
+                    module=args.module,
+                    state_argument=state_argument,
+                    terraform_init_command=terraform_init_command
+            )
+        elif args.subcommand == 'show':
+            if args.plan:
+                state=plan_file
+            else:
+                state=state_file
+
+            cmd = "cd {root_dir}/{terraform_path} && " \
+                  "terraform show {state}".format(
+                    root_dir=self.root_dir,
+                    terraform_path=terraform_path,
+                    state=state
+            )
+        elif args.subcommand == 'import':
+            generate_module_templates = True
+            cmd = "cd {root_dir}/{terraform_path} && " \
+                  "terraform import {state_argument} {vars} module.{module}.{resource} {name}".format(
+                    root_dir=self.root_dir,
+                    command=args.subcommand,
+                    terraform_path=terraform_path,
+                    resource=args.resource,
+                    module=args.module,
+                    name=args.name,
+                    state_argument=state_argument,
+                    vars=vars,
+            )
+        elif args.subcommand == 'console':
+            generate_module_templates = True
+            cmd = "cd {root_dir}/{terraform_path} && " \
+                  "terraform {command} {state_argument} {vars}".format(
+                    root_dir=self.root_dir,
+                    command=args.subcommand,
+                    terraform_path=terraform_path,
+                    state_argument=state_argument,
+                    vars=vars,
+            )
+        elif args.subcommand == 'validate':
+            generate_module_templates = True
+            cmd = "cd {root_dir}/{terraform_path} && {terraform_init_command} " \
+                  "terraform {command} {vars}".format(
+                    command=args.subcommand,
+                    root_dir=self.root_dir,
+                    terraform_init_command=terraform_init_command,
+                    terraform_path=terraform_path,
+                    vars=vars,
+            )
+        elif args.subcommand is not None:
+            # Examples: 
+            #  - command = "state push errored.tfstate"
+            #  - command = "force-unlock <LOCK_ID>"
+            generate_module_templates = True
+            cmd = "cd {root_dir}/{terraform_path} && {terraform_init_command} " \
+                  "terraform {command}".format(
+                    command=args.subcommand,
+                    root_dir=self.root_dir,
+                    terraform_init_command=terraform_init_command,
+                    terraform_path=terraform_path,
+            )
+        else:
+            display('Terraform subcommand \'%s\' not found' % args.subcommand, color='red')
+            return
+
+        if generate_module_templates:
+            self.write_module_templates()
+            post_actions = [self.remove_module_template]
+        else:
+            post_actions = []
+
+        # pass on the terraform args to the terraform command line
+        cmd = ' '.join([cmd] + args.terraform_args + [landscape])
+
+        return dict(
+                command=cmd,
+                post_actions=post_actions
+        )
+
+    def add_profile_vars(self, plan_variables, profile_name):
+        plan_variables['profile'] = '"%s"' % profile_name
+
+        home_dir = os.environ.get('HOME')
+        plan_variables['shared_credentials_file'] = '"{}/.aws/credentials"'.format(home_dir)
+        # plan_variables['access_key'] = '"%s"' % aws.acess_key(profile_name)
+        # plan_variables['secret_key'] = '"%s"' % aws.secret_key(profile_name)
+
+    def get_terraform_path(self):
+        if 'path' in self.cluster_config['terraform']:
+            return self.cluster_config['terraform']['path']
+
+        if 'paths' not in self.cluster_config['terraform']:
+            raise Exception("Could not find 'terraform.path' / 'terraform.paths' in the cluster configuration")
+
+        paths = self.cluster_config['terraform']['paths']
+        selected = self.selected_terraform_path
+        if selected is None:
+            raise Exception('You need to specify which path you want to use with --path-name. Options are: %s ' % paths.keys())
+        
+        try:
+            return paths[selected]
+        except KeyError:
+            raise Exception("Could not find path '%s' in 'terraform.paths'. Options are: %s" % (selected, paths.keys()))
+
+    def get_terraform_src_paths(self):
+        return [self.get_terraform_path()]
+
+    def check_terraform_version(self):
+        expected_version = self.ops_config['terraform.version']
+
+        try:
+            execution = Popen(['terraform', '--version'], stdin=PIPE, stdout=PIPE, stderr=PIPE)
+        except Exception as e:
+            err('Terraform does not appear to be installed, please ensure terraform is in your PATH')
+            raise e
+        current_version, execution_error = execution.communicate()
+        current_version = current_version.replace('Terraform ', '').split('\n', 1)[0]
+        if expected_version == 'latest':
+            return current_version
+
+        if current_version != expected_version and execution.returncode == 0:
+            raise Exception("Terraform should be %s, but you have %s. Please change your version." % (
+                expected_version, current_version))
+
+        return current_version
+
+    def get_templated_files(self):
+        for path in self.get_terraform_src_paths():
+            for source, dest, content in self.template_files(path):
+                yield source, dest, content
+
+    def copy_static_files(self, path, terraform_path):
+        shutil.copytree(os.path.join(self.root_dir, terraform_path), os.path.join(path, terraform_path))
+        shutil.copytree(os.path.join(self.root_dir, 'modules'), os.path.join(path, 'modules'))
+
+    def write_var_file(self, path, variables):
+        fname = os.path.join(path, 'ops.auto.tfvars')
+        with open(fname, 'w') as f:
+            for key, val in variables.items():
+                if val[0] != '"':
+                    val = '"{}"'.format(val)
+                f.write("{key} = {val}\n".format(key=key, val=val))
+
+    def write_module_templates(self, path=''):
+        for original, fname, result in self.get_templated_files():
+            if path:
+                fname = os.path.join(path, fname)
+                folder = os.path.dirname(fname)
+                if not os.path.exists(folder):
+                    os.makedirs(folder)
+            with open(fname, 'w') as f:
+                f.write(result.encode('utf8'))
+
+    def remove_module_template(self):
+        filenames = set()
+        for source, dest, content in self.get_templated_files():
+            filenames.add(dest)
+        for filename in filenames:
+            try:
+                os.remove(filename)
+            except:
+                err('Could not remove file %s' % filename)
+
+    def get_terraform_module_paths(self, rendered):
+        """ Return list of relative module paths that are included in a terraform
+            config file """
+
+        return re.findall('source\s*=\s*"(.+?)"', rendered)
+
+    def template_files(self, path):
+        result = []
+        terraform_file_contents = self.get_terraform_files(path)
+
+        for source in self.list_jinja_templates(path):
+            dest = source.replace(".jinja2", "")
+            config_all = self.cluster_config.all()
+            # Allow access to configuration values in Jinja2. Replace '.' with '_' to make them valid variable names
+            config_all['opsconfig'] = {k.replace('.', '_'): v for k, v in self.ops_config.all().items()}
+            config_all['selected_terraform_path'] = self.selected_terraform_path
+            if config_all.get('terraform', {}).get('boto_profile'):
+                self.add_profile_vars(config_all, config_all['terraform']['boto_profile'])
+            rendered = self.template.render(source, config_all)
+
+            terraform_file_contents.append(rendered)
+
+            result.append((source, dest, rendered))
+
+        # search for module references in all terraform files in this path, including rendered templates
+        for discovered_module in self.find_referenced_modules(path, terraform_file_contents):
+            result.extend(self.template_files(discovered_module))
+
+        return result
+
+    def find_referenced_modules(self, base_path, terraform_files):
+        # look for terraform module references in this path
+        ret = set()
+
+        for rendered in terraform_files:
+            for relative_module_path in self.get_terraform_module_paths(rendered):
+                new_path = os.path.normpath(base_path + '/' + relative_module_path)
+                ret.add(new_path)
+
+        return ret
+
+    def list_files(self, path, extension):
+        template_paths = []
+        loader = FileSystemLoader(path)
+        for fname in loader.list_templates():
+            name, ext = os.path.splitext(fname)
+            template_path = path + '/' + fname
+            # Do not go into terraform community provided modules
+            if ext == extension and '.terraform' not in template_path:
+                template_paths.append(template_path)
+
+        return template_paths
+
+    def get_terraform_files(self, path):
+        ret = []
+        for fname in self.list_files(path, '.tf'):
+            with open(fname) as f:
+                ret.append(f.read())
+
+        return ret
+
+    def list_jinja_templates(self, path):
+        return self.list_files(path, '.jinja2')
+
+    def set_current_working_dir(self):
+        os.chdir(self.root_dir)
